feature: show scanner by default
